[cleanup] remove useless meetme process args.
This was causing conflict with already defined process args (i.e. args
that you pass when you launch the agid). Since it was unused and not
documented, it was removed. Equivalent fonctionality would have been
better provided via a module-specific configuration files (which is more
in line with what some other agid modules do).