make agent client initialization thread safe
It's still possible that the connect method is called more than once,
but this is not a problem since the connect method is thread safe and
does not raise in case the client is already connected.

Still, leave the "connected" test as this is this is the most common case.