Let XiVO use the returned results so as to use LDAP's clever matches
the test "if number in result.itervalues():" was added to the code
on 10 Oct 2013: https://github.com/xivo-pbx/xivo-agid/commit/f51911053285495590ffabf492b8208a01ec583d
for "fix reverse lookup partial patches - Only complete numbers are
matched by a reverse lookup"
according to sebastien_ on IRC, this code was meant to address the
problem that a look-up of "1000" may match "4185551000" because the
number is contained in the string.
The problem with this approach is that the code throws away valid
results returned by clever backends such as LDAP, which return
results for "4185551000" whether LDAP contains "4185551000", "41 8555 1000" or
"41-8555-1000"
So, the correct approach is to let XiVO use the results here, and
address the problem in the way the LDAP filter is constructed, by
removing the wildcards in xivo_dird/directory/data_sources/ldap.py#L77
The other back-ends do not seem to use a wild-card approach, so the
overmatching problem seems only localized to LDAP
I believe that this patch and the one concerning
xivo_dird/directory/data_sources/ldap.py#L77 are addressing the same
issue which commit f51911053285495590ffabf492b8208a01ec583d tried to
address in 2013.