user services: don't check corresponding extension state
The following user services are impacted:

* Do not disturb
* Call recording
* Incoming call filtering
* Forward on non-answer
* Forward on busy
* Unconditional forward

From now on, when these services are enabled for a user, they are always
taken into account when the user emit or receive a call. Previously, the
extension corresponding to the service had to be enabled for the service
to be effectively enabled for a user.