user: wait for mobile without blocking the call
when a user gets called a push mobile event will be sent if the user has a
mobile session, a webrtc line and no active mobile registration.

Instead of calling the interface for that aor,
Local/<aor>@wazo_wait_for_registration will be dialed. All other dialed strings
will ring.

The wait_for_registration dialplan will wait until there is one registration for
that AOR.

If the AOR was already registered with a none mobile session, that session will
start ringing without waiting. In those case, the mobile will have to redirect
the call when it gets woken up.

If there are no active registrations with that AOR the mobile will get waken up
and start ringing.