in_callerid: set callerid using channel variable
Why:
Using the AGI to set the callerid only sets the number and sets the
caller ID name to an empty string. In the dialplan, callerid_forphones
is called next and it uses agi.env['agi_calleridname'] to get the name.
When setting up the env variables, the AGI puts the value "unknown" if
the value does not exist. This makes it so the value "unknown" is set as
the name in the script, which is not the wanted behaviour.