callerid_forphones: remove unknown name condition
Why:
callerid_forphones is executed after in_callerid. We changed the
behaviour of in_callerid so that the number is set as the name if they
were already the same. Previously, in_callerid only set the number and
this erased the name, the AGI substituting it for "unknown" in
callerid_forphones.