recording: fix unittests
the incoming queue and group agi now set the __WAZO_PEER_CALL_RECORD_FILE
variable which is used by the other channel (the callee's) to start the
MixMonitor. XIVO_CALLRECORDFILE is not used to avoid conflicting with the
variable used by the caller's channel. Once on the other channe the
XIVO_CALLRECORDFILE is assigned the value of the WAZO_PEER_CALL_RECORD_FILE to
avoid special cases when starting/stopping