recording: refactor to split filename between 2 variables
the XIVO_CALLRECORDFILE has been replaced by 3 variables

WAZO_CALL_RECORD_SIDE which indicates which side of the call we are
WAZO_CALL_RECORD_FILE_CALLER
WAZO_CALL_RECORD_FILE_CALLEE

This allow the user of both side of the call to control it's own recording.

The __WAZO_LOCAL_CHAN_MATCH variable is used to find the "concrete" channel of a
Local chan in wazo-calld when starting the recording.

When the recording is started from the Queue gosub the channel are not optimized
and the call to the API uses the Local channel. Setting variables on those
channels does not work because the "concrete" channel which is used by the user
do not inherit from the variables after the channel has been created.

Starting the MixMonitor from the AGI does not work because the MixMonitor is
started on the Local channel instead of the "concrete" one. This breaks the
recording status of the wazo-calld call objects because the
WAZO_CALL_RECORD_ACTIVE channel variable is modified on the Local channel after
it has started. The same shenanigans could be done to find the matching channel
but it would have to be done in loop for all calls to GET /calls in calls. Doing
it on record start makes it lazy